==> figures/drug_repurposing/01_check_generated_dataset.ipynb <==
import pandas as pd

####################
# Load the data
####################
train_df = pd.read_csv("complex_disease_train.csv")
valid_df = pd.read_csv("complex_disease_valid.csv")
test_df = pd.read_csv("complex_disease_test.csv")

####################
# Node information
####################
nodes_df = pd.read_csv("data/node.csv", delimiter="\t")
nodes_df = nodes_df.set_index("node_index")
print("--- Nodes")
display(nodes_df.head())
display(nodes_df.shape)

# dictionary to convert node index to node name
node_id2name = nodes_df.set_index(["node_id", "node_type"])["node_name"].to_dict()

####################
# Convert node index to node name
####################
# Relations to be used
relations = ["contraindication", "indication", "off-label use", "rev_contraindication", "rev_indication", "rev_off-label use"]

def validate_id(s):
    if s.endswith(".0"):
        s = s[:-2]
    return s

train_df = train_df[train_df["relation"].isin(relations)].reset_index(drop=True)
train_df["x_id"] = train_df["x_id"].apply(validate_id)
train_df["y_id"] = train_df["y_id"].apply(validate_id)
train_df["x_name"] = train_df.apply(lambda row: node_id2name[(row["x_id"], row["x_type"])], axis=1)
train_df["y_name"] = train_df.apply(lambda row: node_id2name[(row["y_id"], row["y_type"])], axis=1)
print("--- Train data")
display(train_df.head())
display(train_df.shape)

valid_df = valid_df[valid_df["relation"].isin(relations)].reset_index(drop=True)
valid_df["x_id"] = valid_df["x_id"].apply(validate_id)
valid_df["y_id"] = valid_df["y_id"].apply(validate_id)
valid_df["x_name"] = valid_df.apply(lambda row: node_id2name[(row["x_id"], row["x_type"])], axis=1)
valid_df["y_name"] = valid_df.apply(lambda row: node_id2name[(row["y_id"], row["y_type"])], axis=1)
print("--- Valid data")
display(valid_df.head())
display(valid_df.shape)

test_df = test_df[test_df["relation"].isin(relations)].reset_index(drop=True)
test_df["x_id"] = test_df["x_id"].apply(validate_id)
test_df["y_id"] = test_df["y_id"].apply(validate_id)
test_df["x_name"] = test_df.apply(lambda row: node_id2name[(row["x_id"], row["x_type"])], axis=1)
test_df["y_name"] = test_df.apply(lambda row: node_id2name[(row["y_id"], row["y_type"])], axis=1)
print("--- Test data")
display(test_df.head())
display(test_df.shape)

train_df.to_csv("complex_disease_train_w_name.csv", index=False)
valid_df.to_csv("complex_disease_valid_w_name.csv", index=False)
test_df.to_csv("complex_disease_test_w_name.csv", index=False)

train_chemicals = set(train_df[train_df["relation"].apply(lambda x: not x.startswith("rev_"))]["x_name"]) | set(train_df[train_df["relation"].apply(lambda x: x.startswith("rev_"))]["y_name"])
train_diseases = set(train_df[train_df["relation"].apply(lambda x: not x.startswith("rev_"))]["y_name"]) | set(train_df[train_df["relation"].apply(lambda x: x.startswith("rev_"))]["x_name"])

valid_chemicals = set(valid_df[valid_df["relation"].apply(lambda x: not x.startswith("rev_"))]["x_name"]) | set(valid_df[valid_df["relation"].apply(lambda x: x.startswith("rev_"))]["y_name"])
valid_diseases = set(valid_df[valid_df["relation"].apply(lambda x: not x.startswith("rev_"))]["y_name"]) | set(valid_df[valid_df["relation"].apply(lambda x: x.startswith("rev_"))]["x_name"])

test_chemicals = set(test_df[test_df["relation"].apply(lambda x: not x.startswith("rev_"))]["x_name"]) | set(test_df[test_df["relation"].apply(lambda x: x.startswith("rev_"))]["y_name"])
test_diseases = set(test_df[test_df["relation"].apply(lambda x: not x.startswith("rev_"))]["y_name"]) | set(test_df[test_df["relation"].apply(lambda x: x.startswith("rev_"))]["x_name"])

all_chemicals = train_chemicals | valid_chemicals | test_chemicals
print(f"Number of unique chemicals: {len(all_chemicals)} (Train | Valid | Test: {len(train_chemicals)} | {len(valid_chemicals)} | {len(test_chemicals)})")

all_diseases = train_diseases | valid_diseases | test_diseases
print(f"Number of unique diseases: {len(all_diseases)} (Train | Valid | Test: {len(train_diseases)} | {len(valid_diseases)} | {len(test_diseases)})")

with open("all_diseases.txt", "w") as f:
    for disease in all_diseases:
        f.write(f"{disease}\n")

df = pd.concat([train_df, valid_df, test_df], axis=0).reset_index(drop=True)
df = df.drop_duplicates(subset=["x_name", "y_name"])
all_interactions = set(["--".join(sorted([x, y])) for x, y in zip(df["x_name"], df["y_name"])])

density = len(all_interactions) / (len(all_chemicals) * len(all_diseases))

print(f"All interactions: {len(all_interactions)}, Density: {density:.4f}")
